refactor: s3, security, properties, exception 관련 클래스 common 패키지로 이동
