+++ src/main/java/com/thekey/stylekeyserver/common/exception/ApiResponse.java
@@ -1,4 +1,4 @@
-package com.thekey.stylekeyserver.common;
+package com.thekey.stylekeyserver.common.exception;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Getter;
